parse_constrain works. parse_select to improve

+++ include/SQL_parser.h
@@ -57,7 +57,7 @@ typedef enum
 typedef union {
         int id;
         int age;
-        double height;
+        int height;
         char name[MAX_STR_LEN];
 } FieldVal;
